% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foos.R
\name{ar_test}
\alias{ar_test}
\title{AR Test}
\usage{
ar_test(y, X, Z, beta0, intercept = F)
}
\arguments{
\item{y}{The dependent variable of the second stage}

\item{X}{The Endogenous variables}

\item{Z}{The Instruments}

\item{beta0}{The vector of the coefficients to be tested}

\item{intercept}{Logical. Should an intercept be included? Default is false.}
}
\description{
Hypotesis test that do not depends if the instrument is weak or strong
}
\note{
Works with more than one endogenous variable.
If used with the coefficients from TSLS, it computes the J-statistic
}
