% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_foo.R
\name{clr_test}
\alias{clr_test}
\title{CLR Test}
\usage{
clr_test(y, x, Z, H, beta0)
}
\arguments{
\item{y}{The dependent variable of the second stage}

\item{x}{The endogenous variable}

\item{Z}{The matrix of instruments}

\item{H}{The matrix of exogenous variables (not implemented yet)}

\item{beta0}{The hypotesis to be tested}
}
\description{
This is an implementation of Moreira(2003)
}
\note{
Does only work with one endogenous variable
}
