% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foos.R
\name{cragg_donald_stats}
\alias{cragg_donald_stats}
\title{Cragg Donald Statistics}
\usage{
cragg_donald_stats(endogenous, instruments)
}
\arguments{
\item{endogenous}{The endogenous variables, a n by k matrix.}

\item{instruments}{The Instruments, a n by z matrix.}
}
\value{
The Cragg Donald Statistics
}
\description{
Computes de Cragg Donald Statistics
}
\note{
The endogenous variables and instruments should be the columns of the matrix
}
