% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foos.R
\name{stock_yogo_test}
\alias{stock_yogo_test}
\title{Stock and Yogo Test}
\usage{
stock_yogo_test(X, Z, bias = 0.1)
}
\arguments{
\item{X}{The Endogenous variables}

\item{Z}{The Instruments}

\item{bias}{The maximum bias accept. Current values are 0.05,0.1,0.2,0.3. Default is 0.1.}
}
\value{
The value of the Donald Cragg Statistics and the critical value for the number of endogenous and the      number of instruments
}
\description{
Computes the Stock and Yogo test for weak instruments detection, with critical values
}
\section{Warning}{
 Only work until 30 instruments or 3 endogenous variables
}

